+++ CHANGELOG.md
@@ -6,6 +6,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.13.0] - 05-01-2026
+
+### Added
+	- Other input, product only on kanban cards
+
+### Changed
+	- Fix server-api minors
+	- UPDATED data/kanban config
+
 ## [0.11.0] - 04-01-2026
 
 ### Added